Major refactor
Main aspects changed:
Split out the drawing functionality from the Canvas. The canvas now only contains methods to directly manipulate coordinates in a safe way.

Created an Instruction Factory that builds a specific instruction given some input from the parser.
Parser has now been removed, and in it's place there is an instruction factory that builds a specific kind of Instruction based on a given input. If the input is unrecognised, a InvalidInputException is thrown
Each instruction (Line, Rectangle, Fill, Quit, Create) has been extracted into a sub class of Instruction, which exposes only one method executeInstruction().
The Instruction contains only the state (eg Canvas Dimensions), and a set of private methods used to execute that instruction.
This allows for easier extension, simply by creating a class that extends Instruction, and then making sure it can be recognised and built in the factory class.